Fix culling flickering

+++ ivy-graphics/src/base_renderer/mod.rs
@@ -80,4 +80,4 @@ where
     }
 }
 
-type BatchId = usize;
+type BatchId = u32;
